Get ship log images through ShipLogManager
Instead of finding the sprites by traversing the component hierarchy, we
use the ShipLogManager class that can retrieve a ShipLogEntry from its
ID.

This is not only more elegant (and probably faster) but this also fixes
a bug where the suit log couldn't access the sprites unless the ship log
was accessed first. I assume there's some sort of lazy loading going on.